Major refactor of DatabaseEvents code
Clarifying core concepts
------------------------

The refactor started with clarifying core concepts, and giving them better names, e.g.
 * SyncPerformer -> Synchronizer
 * Registry -> EventStore

Single Responsibility
---------------------

Next, I tried to focus on the Single Responsibility Principle and identify the correct
place the code should live. For example, if you look at the old SyncPerformer it knew
a lot about At strings and arrays, which really wasn't it's concern. Compare this to
the new, slimmed down Synchronizer - what it does should be much clearer.

New EventCollection for grouping EventStore#related_events
----------------------------------------------------------

An instance of your Model class can have an #at method specifying a comma-separated At
string. A Clockwork DatabaseEvent::Event is created for each of these separate `At`s.
This means, for a single Model instance, there could be several related events.

Previously, we stored Model events in a Hash keyed by the Model id. If there were multiple
events needed for a Model instance (because it had multiple `At`s) they were stored in an
array inside that hash.

Often we'd need to operate on that array of related events for a particular model instance
(e.g. unregister all of them), but the EventStore (formerly Register) would need to handle
managing that array of related events.

Instead, we've created the concept of an EventCollection, which allows us to perform
group operations in a single go, and provides somewhere to collect methods which act on a
related set of Events.

Miscellaenous Improvements
--------------------------

There were various other small tweaks and improvements along the way.

Conclusion
----------

Again, the code is still far from perfect, but I believe this is a big step in the
right direction, and should be significantly more maintainable.

Squashed Commits
----------------

  * Rename SyncPerformer -> Synchronizer
  * Remove unused DatabaseEvents::Synchronizer SYNCHRONIZERS
  * Refactor DatabaseEvents::Synchronizer#at_strings_for
  * Refactor DatabaseEvents::Synchronizer#create_or_recreate_event
  * Further refactoring of DatabaseEvents::Synchronizer#create_or_recreate_event
  * Extract building the options hash into separate methods.
  * WIP: Beginning to extract EventModelComparer object
  * Refactor DatabaseEvents::Synchronizer#sync
  * WIP: continuing refactor to separate responsibilities
  * Refactor DatabaseEvents::Synchronizer to simplify
  * Rename Registry -> EventStore, and minor refactor
  * Store related Events in EventGroup in EventStore (rather than array)
  * Rename EventGroup -> EventCollection
  * Remove unused DatabaseEvents::EventModelComparer
  * Rename EventStore's @event_groups -> @related_events
  * Use dependency injection to inject manager into EventCollection
  * Change public/private API for EventStore
  * Reorder methods in EventStore
  * Refactor EventCollection #model_ats and #at_strings_for
  * Refactor EventStore#options
  * Include some helpful comments explaining EventStore quirks
  * Fix missing newlines at end of several files