Database-backed events can now set clockwork event :if
Resolves #153, resolves #147 and resolves #113.

An oft-requested feature is to make Database-backed events make use somehow of clockwork's :if option.
For a long time, this seemed impossible to me because the only approach I could envisage was to store Ruby
code in a text field in the database model, and eval it (clearly a bad idea!) However, @adrianpcastro
in #153 suggested an infinitely better approach, that for some reason eluded me: simply having a "if" method
in the database model.

This commit implements that approach, and finally allows Database-backed events to use :if's in the same
way as the standard clockwork events. See the updated README for usage details.

As part of this work:

 * updated README to clarify, and explain usage

 * refactored how we determine whether a model has changed when we sync the models. Previously we checked
   if the name, frequency or at values had changed, and needed some complicated code to compare the model#at
   with those stored in each event created by clockwork, corresponding to a different at time.

   The new approach simply expects models to have an #attributes method (which ActiveRecord models already do)
   which outputs the attribute state as a hash. Now we simply store this, and then check to see if the state
   has changed when we sync.

   Although the public API is the same, this change could potentially cause a subtle change in behaviour in
   terms of syncing events, and for that reason we'll need to bump the major version

 * made the `at` method options for database event models (previously we required it to be defined, and to
   return either nil or '')